requirements.txt
apify-sdk==1.10.0
requests
lxml